feat: add WorkerKind model and refactor Worker to use it
