+++ server/libs/common/src/models/mod.rs
@@ -1,9 +1,11 @@
 mod task_kinds;
 mod tasks;
 mod time_format;
+mod worker_kinds;
 mod workers;
 
 pub use task_kinds::*;
 pub use tasks::*;
 pub use time_format::*;
+pub use worker_kinds::*;
 pub use workers::*;
